authentication/requirements.txt
astroid==2.15.8
blinker==1.6.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
dill==0.3.7
Flask==3.0.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mysqlclient==2.2.0
parso==0.8.3
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
pyserial==3.5
SQLAlchemy==2.0.21
tomlkit==0.12.1
typing_extensions==4.8.0
Werkzeug==3.0.0
wrapt==1.15.0
